requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.12.0
supabase==2.3.4
icalendar==5.0.11
PyPDF2==3.0.1
python-dateutil==2.8.2
httpx==0.26.0
Pillow==10.2.0
